feat: separate components
