+++ src/utils/errors.ts
@@ -1,0 +1,9 @@
+import { Platform } from 'react-native';
+
+const LINKING_ERROR =
+  `The package 'mistica-react-native' doesn't seem to be linked. Make sure: \n\n` +
+  Platform.select({ ios: "- You have run 'pod install'\n", default: '' }) +
+  '- You rebuilt the app after installing the package\n' +
+  '- You are not using Expo Go\n';
+
+export { LINKING_ERROR };
